+++ src/components/Feature/OARemoteOperation.vue
@@ -1,10 +1,14 @@
 <script setup lang="ts">
 import type { OperationSlot } from '../../types'
 import OAContextProvider from '../Context/OAContextProvider.vue'
-import OASpecContent from '../Feature/OASpecContent.vue'
+import OAOperationContent from './OAOperationContent.vue'
 
 const props = defineProps({
   specUrl: {
+    type: String,
+    required: true,
+  },
+  operationId: {
     type: String,
     required: true,
   },
@@ -20,9 +24,10 @@ const slots = defineSlots<Record<string, OperationSlot>>()
 <template>
   <OAContextProvider :spec-url="props.specUrl" @update:spec="emits('update:spec', $event)">
     <template #default="{ openapi }">
-      <OASpecContent
+      <OAOperationContent
         v-bind="{
           ...$attrs,
+          operationId: props.operationId,
           openapi,
         }"
       >
@@ -30,7 +35,7 @@ const slots = defineSlots<Record<string, OperationSlot>>()
         <template v-for="(_, name) in slots" #[name]="slotProps">
           <slot :name="name" v-bind="slotProps || {}" />
         </template>
-      </OASpecContent>
+      </OAOperationContent>
     </template>
   </OAContextProvider>
 </template>
