+++ src/components/index.ts
@@ -1,12 +1,12 @@
-import OALocaleSelect from './Common/OALocaleSelect.vue'
 import OAInfo from './Feature/OAInfo.vue'
 import OAIntroduction from './Feature/OAIntroduction.vue'
+import OALocaleSelect from './Feature/OALocaleSelect.vue'
 import OAOperation from './Feature/OAOperation.vue'
+import OARemoteOperation from './Feature/OARemoteOperation.vue'
+import OARemoteSpec from './Feature/OARemoteSpec.vue'
 import OAServers from './Feature/OAServers.vue'
 import OASpec from './Feature/OASpec.vue'
-import OARemoteOperation from './Remote/OARemoteOperation.vue'
-import OARemoteSpec from './Remote/OARemoteSpec.vue'
-import OASpecSkeleton from './Spec/OASpecSkeleton.vue'
+import OASpecSkeleton from './Feature/OASpecSkeleton.vue'
 
 export {
   OAInfo,
